refactor(bot): :recycle: Update command file structure
imports are one level deep now
