+++ apps/bot/src/commands/nft/index.ts
